requirements.txt
# Core demo libs (CPU oriented)
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
torch>=2.0.0
langchain>=0.0.300
uvicorn
fastapi
pydantic
